+++ src/vec3.rs
@@ -4,24 +4,24 @@ use std::ops::{Add, AddAssign, Div, DivAssign, Index, IndexMut, Mul, MulAssign, 
 
 #[derive(PartialEq, Copy, Clone, Default, Debug)]
 pub struct Vec3 {
-    e: [f64; 3],
+    e: [f32; 3],
 }
 
 impl Vec3 {
     #[inline]
-    pub fn new(x: f64, y: f64, z: f64) -> Self {
+    pub fn new(x: f32, y: f32, z: f32) -> Self {
         Self { e: [x, y, z] }
     }
 
     #[inline]
     pub fn newi(x: i32, y: i32, z: i32) -> Self {
         Self {
-            e: [x as f64, y as f64, z as f64],
+            e: [x as f32, y as f32, z as f32],
         }
     }
 
     #[inline]
-    pub fn fill(t: f64) -> Self {
+    pub fn fill(t: f32) -> Self {
         Vec3 { e: [t, t, t] }
     }
 
@@ -42,7 +42,7 @@ impl Vec3 {
     }
 
     #[inline]
-    pub fn random_range(min: f64, max: f64) -> Self {
+    pub fn random_range(min: f32, max: f32) -> Self {
         Self {
             e: [
                 common::random_double_range(min, max),
@@ -53,43 +53,43 @@ impl Vec3 {
     }
 
     #[inline]
-    pub fn axis(&self, n: usize) -> f64 {
+    pub fn axis(&self, n: usize) -> f32 {
         self.e[n]
     }
 
     #[inline]
-    pub fn x(&self) -> f64 {
+    pub fn x(&self) -> f32 {
         self.e[0]
     }
 
     #[inline]
-    pub fn y(&self) -> f64 {
+    pub fn y(&self) -> f32 {
         self.e[1]
     }
 
     #[inline]
-    pub fn z(&self) -> f64 {
+    pub fn z(&self) -> f32 {
         self.e[2]
     }
 
     #[inline]
-    pub fn length(&self) -> f64 {
-        f64::sqrt(self.length_squared())
+    pub fn length(&self) -> f32 {
+        f32::sqrt(self.length_squared())
     }
 
     #[inline]
-    pub fn length_squared(&self) -> f64 {
+    pub fn length_squared(&self) -> f32 {
         (self.e[0] * self.e[0]) + (self.e[1] * self.e[1]) + (self.e[2] * self.e[2])
     }
 
     #[inline]
     pub fn near_zero(&self) -> bool {
-        const EPS: f64 = 1.0e-8;
+        const EPS: f32 = 1.0e-8;
         // Return true if the vector is close to zero in all dimensions.
-        f64::abs(self.e[0]) < EPS && f64::abs(self.e[1]) < EPS && f64::abs(self.e[2]) < EPS
+        f32::abs(self.e[0]) < EPS && f32::abs(self.e[1]) < EPS && f32::abs(self.e[2]) < EPS
     }
     #[inline]
-    pub fn dot(&self, v: Vec3) -> f64 {
+    pub fn dot(&self, v: Vec3) -> f32 {
         (self.e[0] * v.e[0]) + (self.e[1] * v.e[1]) + (self.e[2] * v.e[2])
     }
 }
@@ -122,24 +122,24 @@ impl AddAssign for Vec3 {
     }
 }
 
-// Vec3 *= f64
-impl MulAssign<f64> for Vec3 {
+// Vec3 *= f32
+impl MulAssign<f32> for Vec3 {
     #[inline]
-    fn mul_assign(&mut self, t: f64) {
+    fn mul_assign(&mut self, t: f32) {
         *self = *self * t;
     }
 }
 
-// Vec3 /= f64
-impl DivAssign<f64> for Vec3 {
+// Vec3 /= f32
+impl DivAssign<f32> for Vec3 {
     #[inline]
-    fn div_assign(&mut self, t: f64) {
+    fn div_assign(&mut self, t: f32) {
         *self = *self / t;
     }
 }
 
 impl Index<usize> for Vec3 {
-    type Output = f64;
+    type Output = f32;
     #[inline]
     fn index(&self, index: usize) -> &Self::Output {
         &self.e[index]
@@ -183,8 +183,8 @@ impl Mul for Vec3 {
     }
 }
 
-// f64 * Vec3
-impl Mul<Vec3> for f64 {
+// f32 * Vec3
+impl Mul<Vec3> for f32 {
     type Output = Vec3;
 
     #[inline]
@@ -193,28 +193,28 @@ impl Mul<Vec3> for f64 {
     }
 }
 
-// Vec3 * f64
-impl Mul<f64> for Vec3 {
+// Vec3 * f32
+impl Mul<f32> for Vec3 {
     type Output = Vec3;
 
     #[inline]
-    fn mul(self, t: f64) -> Vec3 {
+    fn mul(self, t: f32) -> Vec3 {
         Vec3::new(self.x() * t, self.y() * t, self.z() * t)
     }
 }
 
-// Vec3 / f64
-impl Div<f64> for Vec3 {
+// Vec3 / f32
+impl Div<f32> for Vec3 {
     type Output = Vec3;
 
     #[inline]
-    fn div(self, t: f64) -> Vec3 {
+    fn div(self, t: f32) -> Vec3 {
         Vec3::new(self.x() / t, self.y() / t, self.z() / t)
     }
 }
 
-// f64 / Vec3
-impl Div<Vec3> for f64 {
+// f32 / Vec3
+impl Div<Vec3> for f32 {
     type Output = Vec3;
 
     #[inline]
@@ -224,7 +224,7 @@ impl Div<Vec3> for f64 {
 }
 
 #[inline]
-pub fn dot(u: Vec3, v: Vec3) -> f64 {
+pub fn dot(u: Vec3, v: Vec3) -> f32 {
     (u.e[0] * v.e[0]) + (u.e[1] * v.e[1]) + (u.e[2] * v.e[2])
 }
 
@@ -282,7 +282,7 @@ pub fn reflect(v: Vec3, n: Vec3) -> Vec3 {
 /// refractive index if the ray is exiting the material, or its reciprocal if
 /// it's entering.
 #[inline]
-pub fn refract(v: Vec3, n: Vec3, ni_over_nt: f64) -> Option<Vec3> {
+pub fn refract(v: Vec3, n: Vec3, ni_over_nt: f32) -> Option<Vec3> {
     let uv = unit_vector(v);
     let dt = uv.dot(n);
     let discriminant = 1.0 - ni_over_nt * ni_over_nt * (1. - dt * dt);
